requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
deepl==1.8.0
dill==0.3.5.1
idna==3.3
loguru==0.6.0
lxml==4.9.0
multiprocess==0.70.13
mysql-connector==2.2.9
pathos==0.2.9
pox==0.3.1
ppft==1.7.6.5
PyExecJS==1.5.1
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
translators==5.2.2
urllib3==1.26.9
